ON-24665 # Fixed legacy default value files not parsing as private attachments
