+++ jest.config.json
@@ -1,10 +1,23 @@
 {
   "preset": "ts-jest",
   "testEnvironment": "jsdom",
-  "moduleDirectories": ["src", "node_modules"],
-  "roots": [".", "src"],
+  "moduleDirectories": [
+    "src",
+    "node_modules"
+  ],
+  "roots": [
+    ".",
+    "src"
+  ],
   "transform": {
     "^.+\\.jsx?$": "babel-jest"
   },
-  "transformIgnorePatterns": ["/node_modules/(?!@oneblink/apps)"]
+  "transformIgnorePatterns": [
+    "/node_modules/(?!@oneblink/apps)"
+  ],
+  "globals": {
+    "ts-jest": {
+      "isolatedModules": true
+    }
+  }
 }
