+++ package.json
@@ -8,23 +8,23 @@
   },
   "dependencies": {
     "@blinkmobile/geolocation": "^2.0.3",
-    "@emotion/react": "^11.9.0",
-    "@emotion/styled": "^11.8.1",
-    "@mui/icons-material": "^5.6.2",
-    "@oneblink/sdk-core": "^0.4.2-beta.2",
-    "@react-google-maps/api": "2.7.0",
+    "@emotion/react": "^11.10.0",
+    "@emotion/styled": "^11.10.0",
+    "@mui/icons-material": "^5.8.4",
+    "@oneblink/sdk-core": "^0.4.2",
+    "@react-google-maps/api": "2.12.1",
     "blueimp-load-image": "^5.16.0",
     "bulma": "^0.9.3",
     "bulma-slider": "2.0.4",
     "bulma-steps": "^2.2.1",
     "bulma-toast": "^2.4.1",
-    "clsx": "^1.1.1",
-    "color": "^3.2.1",
-    "copy-to-clipboard": "^3.3.1",
-    "date-fns": "^2.28.0",
+    "clsx": "^1.2.1",
+    "color": "^4.2.3",
+    "copy-to-clipboard": "^3.3.2",
+    "date-fns": "^2.29.2",
     "escape-string-regexp": "^5.0.0",
     "file-saver": "^2.0.5",
-    "flatpickr": "^4.6.11",
+    "flatpickr": "^4.6.13",
     "jsqr": "^1.4.0",
     "lodash.clonedeep": "^4.5.0",
     "lodash.debounce": "^4.0.8",
@@ -34,57 +34,56 @@
     "query-string": "^7.1.1",
     "react-google-recaptcha": "^2.1.0",
     "react-input-mask": "^2.0.4",
-    "react-signature-canvas": "^1.0.5",
-    "react-table": "^7.7.0",
-    "sanitize-html": "^2.7.0",
-    "signature_pad": "^3.0.0-beta.4",
+    "react-signature-canvas": "^1.0.6",
+    "react-table": "^7.8.0",
+    "sanitize-html": "^2.7.1",
     "uuid": "^8.3.2",
     "validate.js": "^0.13.1",
     "xml2js": "^0.4.23"
   },
   "devDependencies": {
-    "@babel/preset-env": "^7.16.11",
-    "@mui/lab": "^5.0.0-alpha.79",
-    "@mui/material": "^5.6.1",
-    "@mui/x-date-pickers": "^5.0.0-alpha.2",
+    "@babel/preset-env": "^7.18.10",
+    "@mui/lab": "^5.0.0-alpha.95",
+    "@mui/material": "^5.10.1",
+    "@mui/x-date-pickers": "^5.0.0-beta.6",
     "@oneblink/apps": "^2.0.0-beta.1",
     "@oneblink/release-cli": "^1.0.0",
     "@oneblink/types": "github:oneblink/types",
-    "@types/blueimp-load-image": "^5.14.4",
+    "@types/blueimp-load-image": "^5.16.0",
     "@types/color": "^3.0.3",
     "@types/file-saver": "^2.0.5",
     "@types/filesystem": "^0.0.32",
     "@types/jest": "^27.4.1",
-    "@types/lodash.clonedeep": "^4.5.6",
-    "@types/lodash.debounce": "^4.0.6",
-    "@types/lodash.throttle": "^4.1.6",
+    "@types/lodash.clonedeep": "^4.5.7",
+    "@types/lodash.debounce": "^4.0.7",
+    "@types/lodash.throttle": "^4.1.7",
     "@types/react": "^17.0.43",
     "@types/react-google-recaptcha": "^2.1.5",
-    "@types/react-input-mask": "^2.0.4",
+    "@types/react-input-mask": "^3.0.1",
     "@types/react-router-dom": "^5.3.3",
     "@types/react-signature-canvas": "^1.0.2",
-    "@types/react-table": "^7.7.11",
+    "@types/react-table": "^7.7.12",
     "@types/sanitize-html": "^2.6.2",
     "@types/uuid": "^8.3.4",
-    "@types/xml2js": "^0.4.9",
-    "@typescript-eslint/eslint-plugin": "^5.19.0",
-    "@typescript-eslint/parser": "^5.19.0",
+    "@types/xml2js": "^0.4.11",
+    "@typescript-eslint/eslint-plugin": "^5.33.1",
+    "@typescript-eslint/parser": "^5.33.1",
     "copyfiles": "^2.4.1",
-    "eslint": "^8.13.0",
+    "eslint": "^8.22.0",
     "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "^7.29.4",
-    "eslint-plugin-react-hooks": "^4.4.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-react": "^7.30.1",
+    "eslint-plugin-react-hooks": "^4.6.0",
     "fixpack": "^4.0.0",
     "jest": "^27.5.1",
-    "prettier": "^2.6.2",
+    "prettier": "^2.7.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-router-dom": "^5.3.0",
     "sass": "^1.50.0",
     "ts-jest": "^27.1.4",
     "tsc-watch": "^5.0.3",
-    "typescript": "^4.6.3"
+    "typescript": "^4.7.4"
   },
   "engines": {
     "node": ">=16",
