+++ src/components/formStore/table/ColumnFilters.tsx
@@ -146,6 +146,7 @@ function ColumnFilters({ filter }: Props) {
           <Grid item xs={6}>
             <V4CompatibleDatePicker
               label="After"
+              // @ts-expect-error ???
               showTodayButton
               clearable
               margin="dense"
@@ -181,6 +182,7 @@ function ColumnFilters({ filter }: Props) {
           <Grid item xs={6}>
             <V4CompatibleDatePicker
               label="Before"
+              // @ts-expect-error ???
               showTodayButton
               clearable
               margin="dense"
@@ -222,6 +224,7 @@ function ColumnFilters({ filter }: Props) {
           <Grid item xs={6}>
             <V4CompatibleDateTimePicker
               label="After"
+              // @ts-expect-error ???
               showTodayButton
               clearable
               margin="dense"
@@ -255,6 +258,7 @@ function ColumnFilters({ filter }: Props) {
           <Grid item xs={6}>
             <V4CompatibleDateTimePicker
               label="Before"
+              // @ts-expect-error ???
               showTodayButton
               clearable
               margin="dense"
