+++ src/form-elements/FormElementABN.tsx
@@ -205,6 +205,7 @@ function FormElementABN({
           >
             <InputMask
               mask={isFocused || value ? '99 999 999 999' : '999 999 999'}
+              // @ts-expect-error "react-input-mask" types do not match code
               maskChar=" "
               type="text"
               id={id}
